+++ packages/nns/candid/ledger.d.ts
@@ -1,4 +1,6 @@
+import type { ActorMethod } from "@dfinity/agent";
 import type { Principal } from "@dfinity/principal";
+
 export interface AccountBalanceArgs {
   account: AccountIdentifier;
 }
@@ -44,9 +46,7 @@ export type QueryArchiveError =
       };
     }
   | { Other: { error_message: string; error_code: bigint } };
-export type QueryArchiveFn = (
-  arg_0: GetBlocksArgs
-) => Promise<QueryArchiveResult>;
+export type QueryArchiveFn = ActorMethod<[GetBlocksArgs], QueryArchiveResult>;
 export type QueryArchiveResult =
   | { Ok: BlockRange }
   | { Err: QueryArchiveError };
@@ -95,12 +95,12 @@ export interface TransferFee {
 export type TransferFeeArg = {};
 export type TransferResult = { Ok: BlockIndex } | { Err: TransferError };
 export interface _SERVICE {
-  account_balance: (arg_0: AccountBalanceArgs) => Promise<Tokens>;
-  archives: () => Promise<Archives>;
-  decimals: () => Promise<{ decimals: number }>;
-  name: () => Promise<{ name: string }>;
-  query_blocks: (arg_0: GetBlocksArgs) => Promise<QueryBlocksResponse>;
-  symbol: () => Promise<{ symbol: string }>;
-  transfer: (arg_0: TransferArgs) => Promise<TransferResult>;
-  transfer_fee: (arg_0: TransferFeeArg) => Promise<TransferFee>;
+  account_balance: ActorMethod<[AccountBalanceArgs], Tokens>;
+  archives: ActorMethod<[], Archives>;
+  decimals: ActorMethod<[], { decimals: number }>;
+  name: ActorMethod<[], { name: string }>;
+  query_blocks: ActorMethod<[GetBlocksArgs], QueryBlocksResponse>;
+  symbol: ActorMethod<[], { symbol: string }>;
+  transfer: ActorMethod<[TransferArgs], TransferResult>;
+  transfer_fee: ActorMethod<[TransferFeeArg], TransferFee>;
 }