+++ packages/nns/candid/sns_wasm.d.ts
@@ -1,10 +1,6 @@
+import type { ActorMethod } from "@dfinity/agent";
 import type { Principal } from "@dfinity/principal";
-export interface AddWasmError {
-  error: string;
-}
-export interface AddWasmOk {
-  hash: Array<number>;
-}
+
 export interface AddWasmRequest {
   hash: Array<number>;
   wasm: [] | [SnsWasm];
@@ -20,6 +16,7 @@ export interface DeployNewSnsRequest {
 }
 export interface DeployNewSnsResponse {
   subnet_id: [] | [Principal];
+  error: [] | [SnsWasmError];
   canisters: [] | [SnsCanisterIds];
 }
 export interface DeployedSns {
@@ -56,7 +53,7 @@ export interface NeuronDistribution {
   controller: [] | [Principal];
   stake_e8s: bigint;
 }
-export type Result = { Ok: AddWasmOk } | { Error: AddWasmError };
+export type Result = { Error: SnsWasmError } | { Hash: Array<number> };
 export interface SnsCanisterIds {
   root: [] | [Principal];
   swap: [] | [Principal];
@@ -89,6 +86,9 @@ export interface SnsWasm {
 export interface SnsWasmCanisterInitPayload {
   sns_subnet_ids: Array<Principal>;
 }
+export interface SnsWasmError {
+  message: string;
+}
 export interface SwapDistribution {
   total_e8s: bigint;
   initial_swap_amount_e8s: bigint;
@@ -97,11 +97,12 @@ export interface TreasuryDistribution {
   total_e8s: bigint;
 }
 export interface _SERVICE {
-  add_wasm: (arg_0: AddWasmRequest) => Promise<AddWasmResponse>;
-  deploy_new_sns: (arg_0: DeployNewSnsRequest) => Promise<DeployNewSnsResponse>;
-  get_next_sns_version: (
-    arg_0: GetNextSnsVersionRequest
-  ) => Promise<GetNextSnsVersionResponse>;
-  get_wasm: (arg_0: GetWasmRequest) => Promise<GetWasmResponse>;
-  list_deployed_snses: (arg_0: {}) => Promise<ListDeployedSnsesResponse>;
+  add_wasm: ActorMethod<[AddWasmRequest], AddWasmResponse>;
+  deploy_new_sns: ActorMethod<[DeployNewSnsRequest], DeployNewSnsResponse>;
+  get_next_sns_version: ActorMethod<
+    [GetNextSnsVersionRequest],
+    GetNextSnsVersionResponse
+  >;
+  get_wasm: ActorMethod<[GetWasmRequest], GetWasmResponse>;
+  list_deployed_snses: ActorMethod<[{}], ListDeployedSnsesResponse>;
 }