+++ packages/sns/candid/sns_swap.d.ts
@@ -1,3 +1,5 @@
+import type { ActorMethod } from "@dfinity/agent";
+
 export interface BuyerState {
   icp_disbursing: boolean;
   amount_sns_e8s: bigint;
@@ -71,11 +73,11 @@ export interface TimeWindow {
   end_timestamp_seconds: bigint;
 }
 export interface _SERVICE {
-  error_refund_icp: (arg_0: ErrorRefundIcpRequest) => Promise<{}>;
-  finalize_swap: (arg_0: {}) => Promise<FinalizeSwapResponse>;
-  get_canister_status: (arg_0: {}) => Promise<GetCanisterStatusResponse>;
-  get_state: (arg_0: {}) => Promise<GetStateResponse>;
-  refresh_buyer_tokens: (arg_0: RefreshBuyerTokensRequest) => Promise<{}>;
-  refresh_sns_tokens: (arg_0: {}) => Promise<{}>;
-  set_open_time_window: (arg_0: SetOpenTimeWindowRequest) => Promise<{}>;
+  error_refund_icp: ActorMethod<[ErrorRefundIcpRequest], {}>;
+  finalize_swap: ActorMethod<[{}], FinalizeSwapResponse>;
+  get_canister_status: ActorMethod<[{}], GetCanisterStatusResponse>;
+  get_state: ActorMethod<[{}], GetStateResponse>;
+  refresh_buyer_tokens: ActorMethod<[RefreshBuyerTokensRequest], {}>;
+  refresh_sns_tokens: ActorMethod<[{}], {}>;
+  set_open_time_window: ActorMethod<[SetOpenTimeWindowRequest], {}>;
 }